05_hackathon/requirements.txt
# Dashboard dependencies for Posit Connect Cloud (Streamlit)
# See DEPLOY_POSIT_CONNECT.md for deployment steps.
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
